requirements.txt
mergin-client
geopandas
pandas
google-generativeai
google-api-python-client
google-auth
oauth2client
pyogrio